Create a separated .NET lib for doubly linked list
- Separate console project (demo) & a data struct
- Migrate projects from .NET Framework 3.5 to .NET 6
